speed up subproc init with lazy imports
test suite needed a lot of love to see any benefit
